+++ shell-integration/bash/kitty.bash
@@ -209,9 +209,7 @@ _ksi_main() {
         _ksi_prompt[ps0]+="\[\e]133;C\a\]"
     fi
 
-    [[ -z "${KITTY_KITTEN_EXE}" ]] && KITTY_KITTEN_EXE="kitten"
-    alias edit-in-kitty="${KITTY_KITTEN_EXE} edit-in-kitty"
-    builtin unset KITTY_KITTEN_EXE
+    alias edit-in-kitty="kitten edit-in-kitty"
 
     if [[ "${_ksi_prompt[complete]}" == "y" ]]; then
         _ksi_completions() {